requirements.txt
livekit-agents[openai, silero, elevenlabs]==1.0.11
fastapi==0.115.8
uvicorn==0.34.0
python-dotenv==1.0.1
httpx==0.28.1
ipython==8.13.2
